Create BackendServiceHelper.
